requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9

pandas==2.2.1
numpy==1.26.4
scikit-learn==1.7.2
joblib==1.3.2

opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.2.0

langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1

openai==1.30.1
groq==0.9.0

tqdm==4.66.4
matplotlib==3.8.4
seaborn==0.13.2

requests==2.31.0
python-telegram-bot==20.8